requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
azure-ai-formrecognizer==3.3.2
azure-ai-textanalytics==5.3.0
Pillow==10.1.0
numpy==1.26.2
openai==1.3.7
Werkzeug==3.0.1
